Fix: Moved csv router to router.py
